FlashLights !

+++ Assets/Scripts/IEnemy.cs
@@ -1,9 +1,9 @@
 public interface IEnemy
 {
-    protected int health { get; }
+    protected float health { get; }
     protected string name { get; }
-    protected int damage { get; }
+    protected float damage { get; }
     protected int speed { get; }
-    public void TakeDamage(int damageveil);
-    public void TakeVeil(int damage);
+    public void TakeDamage(float damageveil);
+    public void TakeVeil(float damage);
 }
